Remove TASK_PROTECTED and add concept of "task duration".
A task's duration says that it is allowed to run until some
event occurs.  Then, all tasks which are scoped to that
event are automatically killed.  This is a safer method
than the older TASK_PROTECTED flag.

By default, tasks are scoped to the BALL.  That is, at
endball, all tasks will be killed.  The creator can change
the scope by using one of the newer task_create_gid_while()
or similar, or task_set_duration().  The task itself can
change its duration using the task_remove_duration() call.

The function task_duration_expire() signals an event and
triggers expiry.  The caller should not be scoped to the
event that it is signalling -- trying to kill itself will
cause a fatal error.

So far, valid durations are infinite, live, ball, game, and test.
Live, ball, and game are subsets of each other (in that order).
So signalling game also signals live and ball -- see their
definitions.

All uses of TASK_PROTECTED have been removed and changed
to use, at wost TASK_DURATION_INF which has the same effect,
or better yet something more scoped when possible.
The mode start macro in particular uses TASK_DURATION_BALL.

Tested somewhat in simulation, and stressbuilt OK.