1-color frames are now displayed at 66% intensity by default.
Previously, we mapped the plane into both dark/bright pages,
producing 100% intensity.  This was too much for a real DMD.
Now, the plane is mapped only into the bright phase, and
a blank DMD page is held during the 33% dark phase.  This
looks reasonably well.

This requires reserving 2 blank pages at all times.

One side effect is that even 1-color frames appear to
use two pages now (one is always blank).  The transition
code was updated to deal with this, and is in fact simpler
now.

A new RTT for page flipping is also installed that uses
a lookup table, allowing for different "palettes" in the
future.