Add a new data section called "permanent".
This is for data that is not checksummed nor protected read-only,
but is kept persistently across reboots.

So far, the scores, the player count, and the random seed
have been moved here.

The startup code is changed not to do memory testing on this
area.

Use the __permanent__ prefix on data items that need to be
located here.