Fix two possible lockups in switch scanning:
- If idle did not run for a VERY long time, the counter
could overflow and cause switch queue errors.  Now,
if it is more than 100ms between scan calls, we pretend
it was 100ms since the last call.  This is enough to
prevent the overflow condition.

- Before each switch is serviced, we mark that
dispatching is OK.  Lockups were seen under pinmame
when the system load was heavy, because of all the
switches that got detected on the first scan.