Initial checkin of new image loading procedure.
'imgld' is the utility for generating flat files containing image
content.  The main Makefile can run this, given an input file
defined in the machine Makefile, and paste the output into the
final ROM image.  It also generates a .h file with definitions that
code can use to access the images.

Limitations at present:
- Images are not compressed at all.
- No headers on each image to denote compression type.
- Base page is fixed at 0x20.
- Reserved byte at the beginning of each page is clobbered.
- Image data is not part of the ROM checksum.