Simplify flashing lamp synchronization.
Before, turning on a flashing lamp would force an update to
every other flashing lamp to be on.  Now, the opposite is done;
the state of the newly flashing lamp is synchronized to match
whatever any/all of the other flashing lamps are currently doing.
It is much faster, and reduces some flicker induced if many
lamp flash changes are done consecutively.