TZ: Fix a display bug in the Powerball detected effect.
The initial scroll-in mucks with the way that pages are mapped,
and so the followon flash was not starting in a state that it
expected.  This may need to be fixed permanently in the
transition code, since it used to work OK and may be too
complex for individual effects to have to deal with.