Enable default ballsave lamp effect on all games.
Previously, each machine had to define its own lamp effect
for ball save.  It was optional to do so, although it would be
confusing not knowing how much time was left.

Now, the system ball save leff is available to all games.  It
will use the lamp defined by MACHINE_BALL_SAVE_LAMP if that
is set in the machine config; if not, it will default to
MACHINE_SHOOT_AGAIN_LAMP.

The ball save effects defined in the TZ and WCS configs have
been dropped now that this is provided by the common config.