Simplify the way that all effects are periodically refreshed.
Previously, display, lamp, and music background effects were
all handled separately.  In general, when one of these needs to
be changed, the others do too.

Now, a single periodic function checks for the need to do
a refresh, and updates everything at once.  Because there is a
lot to do, the update occurs in a separate task context.
Presently, the automatic update occurs every 500ms.

Code can force an update faster by calling effect_update_request().
The update will take place during the next 100ms interval, and
will reset the timer so that the next update occurs after a new
500ms interval.  This all cuts down on the frequency of updates
but not so much that it is noticeable.

A new event is thrown every 100ms called 'device_update'.
This is intended to update diverters and other devices which
also need to be governed by complex conditions that are
ever-changing.

The 100ms interval debounces change requests sufficiently so
that when multiple conditions change in a short period, not
as much 'back-and-forth' occurs.

Game code can continue to use deff_update() or music_refresh()
as before, but should be changed to call effect_update_request()
over time.