Support normal declarations of flags and timers.
Previously, flags and timers needed to be declared in the machine
description, which emitted a #define for each.  Now, you can declare
flag_t, global_flag_t, and free_timer_id_t variables throughout the
C code as you would for normal variables.  (flag_t has the same
semantics in that it is per-player.)  None of the user-visible APIs
to manage these changes at all; it is all transparent.

This is preferable over using genmachine because it keeps
the variables closer to their uses, and makes it clear which
modules are responsible for each object.  genmachine is intended
to be used for _machine_ definitions rather than _rules_.
Also, adding/deleting these types of variables can now be done
without recompiling everything (re-running genmachine requires
most C files to be compiled again).

For backwards compatibility, the md file sections are still
supported; entries here will be emitted into a mach-vars.c file
which is compiled as normal C code.  At some point these entries
could be migrated back into the machine game code.

Internally, only the addresses of these variables are used.  They
are not linked into any real sections of the running program.
We just use the linker facility to assign the addresses.  The
conversion to addresses is done internally; for the 6809 it is
optimized using an asm macro to keep it 8-bit only.

There are other types not yet converted to this style which may
be added later: adjustments and audits come to mind.