Fix switch scanning logic that could cause switches to be missed.
Previously, if a switch was marked as needed no long debounce time,
but the scanner did not see it before it deactivated, it was
considered a miss and the handler was not called.  As checking
only happens every 16ms, or possibly longer in a busy system, a
switch that closed and then opened rapidly, say less than 16ms,
would never be handled.

Now, when a switch has zero debounce time, the check for rapid
deactivation is not performed at all.  The handler will always be
called.  The logic to check for deactivation only applies when the
switch needed a much longer debounce.  This turns out to be a rare
case, and so the new logic is actually more optimal.

The minimum switch closure time is 4ms -- up to 2ms for it to
be seen for the first time, and 2ms more for it to be confirmed.