Do not allow GI to be allocate from a shared lamp effect.
Due to the current design of the lamp effect module, GI strings cannot
be controlled by more than 1 lamp effect at a time.  Thus shared leffs
are not allowed to use them.

genmachine will now detect when you try to do this, and fail.

Future enhancements to the system software will allow this.