Native mode: Move periodic processing into realtime handler.
Previously periodic processing was done in the main task when it was
about to exit, but that meant periodics weren't called until the
very end of init.  Normally they would be called during late init,
and some code depends on this to work properly.

Periodics are now called every 16ms from the realtime handler,
which is about how it works on real hardware.