Fix bug in triac update, causing unnecessary writes to the latches.
Note that the rtt reads the values of the dimming arrays based on
the state of the AC cycle (zc_get_timer), whose maximum value was
larger than NUM_BRIGHTNESS_LEVELS.  Thus undefined values were being
read, outside the range of the array.

This change increases the array size to the maximum, so that
array out-of-bounds does not occur.