Add the 'have' and 'nohave' Makefile functions.
This is now the recommended way to turn on/off features from all
of the sub-Makefiles.  It creates both the C/assembly #define
plus a make variable.  Over time lots of code in the Makefiles
should be changed to use this instead of modifying CFLAGS or
making new makefile variables.