Native mode: support automated stress testing.
With these changes, stress mode can now be invoked from the command-line
allowing for totally automated testing.  The current stresstest runs
for 30 minutes then exits with success.

The native mode now exits with nonzero if a fatal occurred.

Use --error-crash option to trigger a coredump on fatal error, so that
it can be analyzed later, rather than just exiting.

Use --late option to load a script after the system has initialized;
this is for switch simulation.  Without it the script parsing will
try to occur too early.

Run 'tools/native_stress' from command-line to start a test.
The option '-' means to test the current build; otherwise, it is a list
of machines to recompile and run.  On error, the test will stop, and
gdb will be started.