Add new code banks exclusively for display & lamp effects.
Usage of this new banks is optional.  Effects must be declared
in a separate .c file and placed into DEFF_OBJS or LEFF_OBJS, and
the effect entry in the .md file must specify DEFF_PAGE or LEFF_PAGE.

The intent is to keep all effect functions localized, so that
libraries of common functions can also be kept in the same bank,
which eliminates the need for lots of far calls, which makes them
run a lot faster.