Initial checkin of the score rank module.
This module must be enabled by defining CONFIG_SCORE_RANK := y in the Makefile
or .config.  When enabled, the system periodically updates a ranking of all
scores in a multiplayer game.  It will generate an event 'score_rank_change'
anytime it changes, at most once per second.  The ranks are guaranteed to be
correct at ball start and end game, too.