Add 'task tail pointer' which tracks top of usable task table.
The task table has a large size to accommodate lots of tasks
running simultaneously, but most times, not all of these task
slots will be used.  API performance often is improved if the
table size is smaller.

We now track top-of-table via a pointer, and only scan that
many entries.  Most of the time, the pointer is kept low.
When more tasks are needed, the pointer is extended; when the
task pressure reduces, the pointer will be moved back again.
Thus overall the performance of the task APIs is improved.