Support template inheritance, using the @@include directive.
This lets you create templates which are specialized versions of
other templates.  For example, slingshot and jet bumper drivers are
just specific versions of the 'spsol' template, with predefined values
for the timings appropriate to that instances.  Machines can just
instantiate the 'sling' or 'jet' driver; the timings are now common
to all instances and don't need to specified completely everytime.

To define a specialized template, two new keywords are introduced.
@@weakdefine is used to set the values for the parameters of the
generic driver.  The machine can still define this as before, overriding
what the specialization says.  @@include should appear last and says
which other template file to include.

The @@class directive is applied only the first time it is seen,
so @@class lines inside an @@included file are ignored.  (For tracking
they are added to a "superclass list" that is printed in the comments in
the generated files.)