Enhance common logic for serving balls to the manual/auto plunger.
- Use new API serve_ball_auto() to autolaunch a ball, rather than
kicking the trough manually.  This is a legal call on manual
plunger games, too, and will default to serve_ball() behavior.

- launch_ball() now has the common logic for autofiring the ball.
Use this instead of kicking the autoplunger manually.  (This API
does the right thing for Twilight Zone, with its quirky autoplunger.)

- Launch ball will retry if the shooter switch continues to be seen
after firing.

- Pressing the launch button (or flippers, if enabled) will now
fire the autolauncher even if the shooter switch is not seen, prior
to valid playfield.

- serve_ball() will flash the launch button lamp if one is defined.
The lamp stops flashing as soon as valid playfield is asserted.

- Timed plunger logic does not require seeing the shooter switch;
it is enabled as part of serve_ball() now.

- When multiple balls need to be autolaunched, there is a minimum
delay between balls now.  Defaults to 3 seconds, but machine can
override it.

- Simplified logic to autofire automatically, without a button press.

FIXUP LATER: Powerball test should still use manual trough kick;
serve_ball() not suitable for test mode?