Rename 'include/sys' to 'include/system'.
This was causing lots of difficulty when compiling in native
mode, because files like include/sys/math.h were conflicting
with the native counterpart like <sys/math.h> in /usr/include.
The include path search probably could be improved, but this
seemed a much easier and safer change.  Mostly just freewpc.h
needed to be changed to use the new names, but a few other files
had to be modified too.