Add CONFIG_MMIO build option.
This option says that readb() and writeb() calls should be
handled as memory-mapped I/O.  That is, the IOPTR argument
to these calls is a memory address, and the calls are translated
to inline reads/writes.

This is a platform dependent option.  When building for
simulation, it should never be defined.