Handle slings/jets with stuck on switches.
Previously, after the solenoid was fired, we just waited a bit
before checking for a switch closure again.  However, this did
not verify that the switch actually opened while the coil was
off.  If the switch was stuck on, it would repeatedly fire.

Now, during the off phase, ensure that the switch opens again
before allowing another activation.

Even if the device is disabled (such as during bonus), we will
continue to poll it, so it will not fire again under any
circumstances until the switch clears.