Rename the native platform to "simulation".
Platforms and simulation are two independent things.  For example,
WPC is a platform, and we can build it with or without simulation.
Also, the source code uses CONFIG_NATIVE to refer to the native _CPU_,
independent of simulation.

So, to make this more clear, move the "platform/native" source tree
into "sim".  Use CONFIG_SIM := y to enable it.  For now, NATIVE := y
will still work, too.