Rewrite most of the flipper code detector.
The concept of a global table of handlers and codes is eliminated.
Modules that want to handle a code should catch the flipper_code_entered
event, which is thrown whenever a new code is locked in.  They should use
the flipcode3() or flipcode4() APIs to retrieve the code just locked.

This is much simpler than before.

Also added this module back into PIC and WPC95 games, which had been
taken out long ago for space savings.