Enable profiling on 6809 systems.
Define CONFIG_PROFILE to enable this feature.  gcc will insert
a call to "_mcount" at the top of every C function.  This
function can be used for a multitude of purposes; at present,
it does nothing unless CONFIG_PROFILE_BPT is set, which ties
the profiling into the breakpoint system.  This can be
expanded later.

Do not define this if you don't really need it, as it only adds
overhead.