Initial implementation of the deff compiler (dcc).
The intent is to write display effects in a new language that is
much simpler than emitting C code.  Depending on flashing,
blinking text, color choices... emitting the right C calls is not
always easy.  The deff compiler's job is to choose the right functions
for you, so you can focus on just saying what you want to happen.

This is not functional yet, but a good chunk of the core of how
it should work is there.  Over time, common display effects should
be migrated to this approach -- separate from the core CPU code --
and the compiler will be updated as necessary to support that.