Move switch/lamp rtts into platform code.
This is how the platform abstraction needs to be defined to work
for a lot of cases.  The details of writing strobes, etc. is too
WPC specific to work for other platforms nearly the same.  Best is
to let each platform write its own interrupt handler entirely.

The kernel files should define the data structures to store state
and provide the user APIs.