Reset the edge bit when a switch changes logical state.
The edge bit is what implements low-level debouncing, and requires
two consecutive readings at the new level.  Once the new level is
latched, we need to restart and wait for two more different levels.

Without this change, it would be possible for a switch to return
to its old level immediately and that would be accepted, without
debouncing.